chore: drop now-redundant `user.id as i64` casts
`arium::User.id` widened from i32 to i64 in Riparion/arium#22 (already
merged and pulled in by Cargo.lock), so the three `user.id as i64`
sites in `server::{mod, admin::mod, comments}` are casting i64 to i64.
Clippy's `unnecessary_cast` lint flagged them.

+++ src/server/admin/mod.rs
@@ -36,7 +36,7 @@ pub(crate) async fn can_edit_post(
         .current_user
         .as_ref()
         .filter(|u| !u.anonymous)
-        .map(|u| u.id as i64)
+        .map(|u| u.id)
         .ok_or_else(|| ServerFnError::new("Not signed in."))?;
     arium_dioxus::require_resource_or_permission(
         authority.0.as_ref(),
